requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
h11==0.14.0
idna==3.10
MouseInfo==0.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.1.0
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
types-requests==2.32.0.20250328
typing_extensions==4.13.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
